+++ windows/flutter/generated_plugins.cmake
@@ -4,6 +4,8 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   file_selector_windows
+  flutter_secure_storage_windows
+  local_auth_windows
   sqlite3_flutter_libs
   url_launcher_windows
 )